Separated the network project and created common project
