+++ AElf.Common/Synchronisation/ReaderWriterLock.cs
@@ -1,37 +1,8 @@
 ﻿using System;
 using System.Threading.Tasks;
 
-namespace AElf.Kernel.Lock
+namespace AElf.Common.Synchronisation
 {
-    /// <summary>
-    /// An async reader writer lock for concurrent and exclusive work.
-    /// </summary>
-    public interface ILock
-    {
-        /// <summary>
-        /// parallel on the default scheduler.
-        /// </summary>
-        /// <param name="func"></param>
-        /// <returns></returns>
-        Task<T> ReadLock<T>(Func<T> func);
-
-        /// <summary>
-        /// Delegates calling this method will be done in sequentially
-        /// </summary>
-        /// <param name="func"></param>
-        /// <returns></returns>
-        Task<T> WriteLock<T>(Func<T> func);
-
-
-        /// <summary>
-        /// for delegate calling without return type
-        /// </summary>
-        /// <param name="action"></param>
-        /// <returns></returns>
-        Task WriteLock(Action action);
-    }
-
-    
     /// <inheritdoc />
     /// <summary>
     /// Initializes a new instance of the object with ability to cancel locked tasks.
@@ -64,6 +35,4 @@ namespace AElf.Kernel.Lock
             return ExclusiveWritrer.StartNew(action);
         }
     }
-    
-    
 }
