+++ AElf.Kernel.Modules.AutofacModule/AElf.Kernel.Modules.AutofacModule.csproj
@@ -4,9 +4,11 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Autofac" Version="4.6.2" />
+    <PackageReference Include="NLog" Version="4.5.6" />
     <PackageReference Include="xunit.frameworks.autofac" Version="0.1.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\AElf.Kernel\AElf.Kernel.csproj" />
+    <ProjectReference Include="..\AElf.Network\AElf.Network.csproj" />
   </ItemGroup>
 </Project>
