+++ AElf.Kernel/Miner/Miner.cs
@@ -1,12 +1,8 @@
-﻿using System.Collections.Generic;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
-using AElf.Kernel.Managers;
 using AElf.Kernel.Services;
 using AElf.Kernel.TxMemPool;
-using Google.Protobuf;
-using ReaderWriterLock = AElf.Kernel.Lock.ReaderWriterLock;
-
+using ReaderWriterLock = AElf.Common.Synchronisation.ReaderWriterLock;
 
 namespace AElf.Kernel.Miner
 {
