+++ AElf.Kernel/TxMemPool/TxPoolService.cs
@@ -3,9 +3,8 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using AElf.Kernel.Managers;
 using AElf.Kernel.Services;
-using ReaderWriterLock = AElf.Kernel.Lock.ReaderWriterLock;
+using ReaderWriterLock = AElf.Common.Synchronisation.ReaderWriterLock;
 
 namespace AElf.Kernel.TxMemPool
 {
