+++ AElf.Launcher/AElf.Launcher.csproj
@@ -9,10 +9,11 @@
       <Name>AElf.Kernel.Modules.AutofacModule</Name>
     </ProjectReference>
     <ProjectReference Include="..\AElf.Kernel\AElf.Kernel.csproj" />
+    <ProjectReference Include="..\AElf.Network\AElf.Network.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Autofac" Version="4.8.0" />
     <PackageReference Include="CommandLineParser" Version="2.2.1" />
-    <PackageReference Include="NLog" Version="4.5.4" />
+    <PackageReference Include="NLog" Version="4.5.6" />
   </ItemGroup>
 </Project>
