+++ AElf.Network/AElf.Network.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Google.Protobuf" Version="3.5.1" />
+    <PackageReference Include="NLog" Version="4.5.6" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\AElf.Common\AElf.Common.csproj" />
+  </ItemGroup>
+</Project>
