+++ AElf.Network/Data/NodeData.cs
