+++ AElf.Network/Helpers/StringHelpers.cs
