+++ AElf.sln
@@ -27,6 +27,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AElf.Database.Tests", "AElf
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AElf.Launcher", "AElf.Launcher\AElf.Launcher.csproj", "{E8D123D9-BE31-4B7B-8109-0D5B3A2FBDDA}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AElf.Network", "AElf.Network\AElf.Network.csproj", "{5FEFB079-F6EB-48AF-97C7-6C582B9BDB8A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AElf.Common", "AElf.Common\AElf.Common.csproj", "{4EC22E9F-7019-470B-B990-B9E020DBBFB2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -81,6 +85,14 @@ Global
 		{E8D123D9-BE31-4B7B-8109-0D5B3A2FBDDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E8D123D9-BE31-4B7B-8109-0D5B3A2FBDDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E8D123D9-BE31-4B7B-8109-0D5B3A2FBDDA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5FEFB079-F6EB-48AF-97C7-6C582B9BDB8A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5FEFB079-F6EB-48AF-97C7-6C582B9BDB8A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5FEFB079-F6EB-48AF-97C7-6C582B9BDB8A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5FEFB079-F6EB-48AF-97C7-6C582B9BDB8A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4EC22E9F-7019-470B-B990-B9E020DBBFB2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4EC22E9F-7019-470B-B990-B9E020DBBFB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4EC22E9F-7019-470B-B990-B9E020DBBFB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4EC22E9F-7019-470B-B990-B9E020DBBFB2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
